refactor: 구조에 맞게 타입 정리 [TM-131]
