+++ src/entities/calendar/calendar.types.ts
@@ -1,4 +1,6 @@
-import { ButtonHTMLAttributes } from 'react';
+// 이벤트 Date 타입
+type ValuePiece = Date | null;
+type Value = ValuePiece | [ValuePiece, ValuePiece];
 
 // 이벤트 타입
 interface CalendarEvent {
@@ -24,23 +26,4 @@ type FetchEventResponse = {
   planDto: PlanDTO;
 };
 
-interface IEventSummaryProps {
-  children: React.ReactNode;
-  onClick?: () => void;
-}
-
-type ActionButtonType = 'edit' | 'register';
-
-interface IActionButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  buttonType?: ActionButtonType;
-  onClick?: () => void;
-  disabled?: boolean;
-}
-
-export type {
-  CalendarEvent,
-  FetchEventResponse,
-  IEventSummaryProps,
-  ActionButtonType,
-  IActionButtonProps,
-};
+export type { Value, CalendarEvent, FetchEventResponse };
