+++ src/features/calendar/calendar.constants.ts
@@ -1,9 +1,12 @@
-const CALENDAR_HEIGHT_DEFAULT = '520px';
-const CALENDAR_HEIGHT_EXTENDED = '595px';
-const CALENDAR_MAX_WEEKS = 6;
+// 보정값을 상수로 분리
+const POPOVER_OFFSET = 66.5;
+const POPOVER_SHIFT_LEFT = '-25%';
+const POPOVER_SHIFT_RIGHT = '25%';
+const POPOVER_CENTERED = '0%';
 
 export {
-  CALENDAR_HEIGHT_DEFAULT,
-  CALENDAR_HEIGHT_EXTENDED,
-  CALENDAR_MAX_WEEKS,
+  POPOVER_OFFSET,
+  POPOVER_SHIFT_LEFT,
+  POPOVER_SHIFT_RIGHT,
+  POPOVER_CENTERED,
 };
