+++ src/features/calendar/calendar.types.ts
@@ -1,0 +1,3 @@
+type EventEditorMode = 'register' | 'edit' | 'read';
+
+export type { EventEditorMode };
