+++ src/features/calendar/model/index.ts
@@ -1,4 +1,3 @@
 export * from './calendarStore';
-export * from './useCalendarViewModel';
 export * from './usePopoverViewModel';
 export * from './useEditorModalViewModel';
