+++ src/widgets/calendar/calendar.constants.ts
@@ -1,0 +1,9 @@
+const CALENDAR_HEIGHT_DEFAULT = '520px';
+const CALENDAR_HEIGHT_EXTENDED = '595px';
+const CALENDAR_MAX_WEEKS = 6;
+
+export {
+  CALENDAR_HEIGHT_DEFAULT,
+  CALENDAR_HEIGHT_EXTENDED,
+  CALENDAR_MAX_WEEKS,
+};
