--- src/widgets/calendar/config/calendar.constants.ts
@@ -1,4 +1,0 @@
-// 보정값을 상수로 분리
-const POPOVER_OFFSET = 66.5;
-
-export { POPOVER_OFFSET };
