+++ src/widgets/calendar/lib/getCalendarHeight.ts
@@ -2,8 +2,8 @@ import {
   CALENDAR_HEIGHT_DEFAULT,
   CALENDAR_HEIGHT_EXTENDED,
   CALENDAR_MAX_WEEKS,
-} from '../../../features/calendar/calendar.constants';
-import { Value } from '@/widgets/calendar/calendar.types';
+} from '@/widgets/calendar/calendar.constants';
+import { Value } from '@/entities/calendar/calendar.types';
 
 // 한 달에 몇 주가 있는지 구하는 함수
 const getWeeksInMonth = (date: Date) => {
