+++ src/widgets/calendar/model/useCalendarViewModel.ts
@@ -1,9 +1,9 @@
 import { useEffect, useMemo, useState } from 'react';
 import { useShallow } from 'zustand/shallow';
-import { Value } from '@/widgets/calendar/calendar.types';
 import dayjs from 'dayjs';
-import { useCalendarStore } from './calendarStore';
+import { useCalendarStore } from '../../../features/calendar/model/calendarStore';
 import { getCalendarHeight } from '@/widgets/calendar/lib/getCalendarHeight';
+import { Value } from '@/entities/calendar/calendar.types';
 
 const useCalendarViewModel = () => {
   const { setSelectedDate } = useCalendarStore(
